Move examples to separate file for simple repository
